Finalización del service Portfolio para envios de datos al lado del cliente

+++ src/main/java/com/nietott/portfolio/DTO/skillsDTO.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @AllArgsConstructor
 @Data
 public class skillsDTO {
-    private int skillId;
-    private int userId;
+    private Long skillId;
+    private Long userId;
     private String skillName;
     private String proficiencyLevel;
     private String icon;
